Update standalone pages for new permalink strategy in Jekyll 0.5.2
As of Jekyll 0.5.2, if the site is configured for "pretty" permalinks
(in _config.yml), then that permalink strategy is applied to all non-index
pages (i.e., all pages not named like "index.*").  For example, by
default a page named "about.md" would be generated as /about/index.html.
To have the page generated as "about.html" (which is certainly my desired
outcome for all non-blog-related pages), you must explicitly declare the
permalink in the yml front matter for the page.

Jekyll 0.5.2 Release Notes:

  http://github.com/mojombo/jekyll/blob/f1c8e388e406e2cc022354562d8bd683d1262072/History.txt

Commit that implemented this change in Jekyll:

  http://github.com/mojombo/jekyll/commit/49c39f43a1c19958ab414692736987b2ae1c8b37